\name{SA24204137-package}
\alias{SA24204137-package}
\alias{SA24204137}
\docType{package}
\title{
  Final Project for 'Statistical Computing' course
}
\description{
  Homework and function are included in this package.
}
\author{
Liu Yao

Maintainer: Liu Yao <2323610037@qq.com>
}
\keyword{ package }
