% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/SA24204137R.R
\name{vif}
\alias{vif}
\title{Calculate Variance Inflation Factor}
\usage{
vif(x)

vif(x)
}
\arguments{
\item{x}{A numeric matrix or data frame of predictor variables (independent variables).}
}
\value{
A numeric vector of VIF values for each predictor variable in the dataset.

A numeric vector of VIF values for each predictor variable.
}
\description{
Calculate Variance Inflation Factor (VIF) for each predictor variable in a dataset.

This function serves as an R wrapper to the C++ implementation of VIF calculation.
}
\examples{
\dontrun{
data(mtcars)
vif(as.matrix(mtcars[, -1]))
}
\dontrun{
data(mtcars)
vif(as.matrix(mtcars[, -1]))
}
}
